Delete branch after PR is closed
https://github.com/codelittinc/roadrunner/issues/11

This is necessary because leaving old branches alive
in github becomes messy fast.

These changes leverage the octokit api to delete a
branch associated with a given closed PR.

In some cases if the request fails we may have some orphaned
branches that slip under the radar. We should keep an eye
out for them